Fixes getServerVersion gql
