+++ pkg/apis/infrastructure/infra.go
@@ -189,3 +189,43 @@ func DisconnectInfra(projectID string, infraID string, cred types.Credentials) (
 		return DisconnectInfraData{}, err
 	}
 }
+
+// GetServerVersion fetches the GQL server version
+func GetServerVersion(endpoint string) (ServerVersionResponse, error) {
+	var gqlReq ServerVersionRequest
+	var err error
+
+	gqlReq.Query = ServerVersionQuery
+	query, err := json.Marshal(gqlReq)
+	if err != nil {
+		return ServerVersionResponse{}, err
+	}
+	resp, err := apis.SendRequest(
+		apis.SendRequestParams{
+			Endpoint: endpoint + utils.GQLAPIPath,
+		},
+		query,
+		string(types.Post),
+	)
+	if err != nil {
+		return ServerVersionResponse{}, err
+	}
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	defer resp.Body.Close()
+	if err != nil {
+		return ServerVersionResponse{}, err
+	}
+	if resp.StatusCode == http.StatusOK {
+		var version ServerVersionResponse
+		err = json.Unmarshal(bodyBytes, &version)
+		if err != nil {
+			return ServerVersionResponse{}, err
+		}
+		if len(version.Errors) > 0 {
+			return ServerVersionResponse{}, errors.New(version.Errors[0].Message)
+		}
+		return version, nil
+	} else {
+		return ServerVersionResponse{}, errors.New(resp.Status)
+	}
+}
